+++ changes.md
@@ -1,6 +1,10 @@
 Dart Web Toolkit Changes
 ========================
 
+**March 27, 2013
+
+* Because bug in Dart http://code.google.com/p/dart/issues/detail?id=9356 the logic of set/get text in ValueBoxBase has been changed.
+
 **March 26, 2013
 
 * Fixed bug:Hyperlink.fromElement generates null pointer exception. Changed Hyperlink.